requirements.txt
requests~=2.28.2
aiohttp~=3.8.4
SQLAlchemy<2.0
python-dotenv~=1.0.0
asyncpg~=0.27.0
